% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_num.R
\name{apa_num}
\alias{apa_num}
\title{Format Numbers to APA Style}
\usage{
apa_num(
  x,
  n_decimal_places = 2,
  remove_leading_zeros = TRUE,
  add_space_start = TRUE,
  na_to_blank = TRUE,
  blank_to_blank = TRUE
)
}
\arguments{
\item{x}{A numeric or character vector. These are the numbers to be formatted.}

\item{n_decimal_places}{An integer specifying the number of decimal places to round to. Default is 2.}

\item{remove_leading_zeros}{A logical indicating whether to remove leading zeros. Default is TRUE.}

\item{add_space_start}{A logical indicating whether to add a space before the decimal point for numbers
less than 1 and greater than -1, without a leading 0 or - sign. Default is TRUE.}

\item{na_to_blank}{A logical indicating whether to convert NA values to blank strings. Default is TRUE.}

\item{blank_to_blank}{A logical indicating whether to convert blank strings to blank strings. Default is TRUE.}
}
\value{
A character vector of APA formatted numbers corresponding to 'x'.
}
\description{
This function formats a given numeric input to APA style. This involves rounding
the number to a specified number of decimal places. Any leading zeros before
the decimal point are also removed. It can optionally add a space before the decimal point for numbers
less than 1 and greater than -1, without a leading 0 or - sign.
}
\details{
Character inputs are converted to numeric using \code{as.numeric} before being formatted.
A warning regarding "NAs introduced by coercion" will be given if any non-numeric values are present in the input.
}
\examples{
apa_num(-0.1) # returns "-.10"
apa_num(-100.12314) # returns "-100.12"
apa_num(0.123456) # returns " .12"
apa_num(0.123456, add_space_start = FALSE) # returns ".12"
apa_num(0.987654, n_decimal_places = 3) # returns " .988"
apa_num(c(-0.123456, 0.987654), n_decimal_places = 2) # returns "-.12" " .99"

}
