% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Create a ggplot2 Correlation Matrix}
\usage{
plot_correlations(
  dat,
  text_size_r = 2.8,
  text_size_axis = 10,
  sample_size = TRUE,
  p_threshold_col = NULL,
  confidence_interval = TRUE,
  low_colour = "#0072B2",
  high_colour = "#D55E00",
  mid_colour = "white",
  abs_colour = TRUE,
  cluster_variables = FALSE,
  n_decimal_places = 2,
  n_decimal_places_r = NULL,
  n_decimal_places_ci = NULL,
  suppress_warning_message = FALSE
)
}
\arguments{
\item{dat}{A dataframe or matrix for which to compute correlations. Note: A correlation matrix should not be provided as input.}

\item{text_size_r}{A scalar that adjusts correlation text size by a specific magnification factor.}

\item{text_size_axis}{A scalar that adjusts axis text size by a specific magnification factor.}

\item{sample_size}{Logical. If TRUE, the function includes sample sizes on the upper diagonal. If FALSE, these are left blank.}

\item{p_threshold_col}{A numeric value indicating the p-value threshold for colouring.}

\item{confidence_interval}{Logical. If TRUE, the function includes confidence intervals on the upper diagonal. If FALSE, these are left blank.}

\item{low_colour}{Hex colour code for the lowest correlation value.}

\item{high_colour}{Hex colour code for the highest correlation value.}

\item{mid_colour}{Hex colour code for the middle point in the correlation scale.}

\item{abs_colour}{Logical. If TRUE, absolute correlation values (i.e., ignoring positive/negative status) are used for determining square colours.}

\item{cluster_variables}{Logical. If TRUE, variables are clustered in the plot. Default is FALSE.}

\item{n_decimal_places}{The number of decimal places to use for plot text.}

\item{n_decimal_places_r}{The number of decimal places to use for correlation (over-rides n_decimal_places)}

\item{n_decimal_places_ci}{The number of decimal places to use for confidence intervals (over-rides n_decimal_places)}

\item{suppress_warning_message}{Logical. If TRUE, suppresses warning message about the function being in beta.}
}
\value{
A ggplot object of the correlation matrix.
}
\description{
The function provides a wrapper around ggplot2 to easily create a correlation matrix.
}
\examples{
X = sapply(1:10, function(i) rnorm(100))
X = as.data.frame(X)

plot_correlations(X, sample_size = TRUE, confidence_interval = FALSE)

plot_correlations(X, sample_size = TRUE, confidence_interval = FALSE) +
ggplot2::labs(title = "My Title")
}
