% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_correlations.R
\name{plot_missing_correlations}
\alias{plot_missing_correlations}
\title{Plot Missing Correlations}
\usage{
plot_missing_correlations(
  dat,
  textadjust = 2,
  p_threshold_col = 0.05,
  low_colour = "#0072B2",
  high_colour = "#D55E00",
  mid_colour = "white",
  abs_colour = TRUE,
  set_colour_to_p_value = FALSE,
  cluster_variables = TRUE,
  n_decimal_places = 3,
  suppress_warning_message = FALSE
)
}
\arguments{
\item{dat}{A dataframe with missing data that you want to analyze.}

\item{textadjust}{Adjust text size by a magnification factor.}

\item{p_threshold_col}{A numeric value indicating the p-value threshold for colouring.}

\item{low_colour}{Hex colour code for the lowest correlation.}

\item{high_colour}{Hex colour code for the highest correlation.}

\item{mid_colour}{Hex colour code for the midpoint of the correlation.}

\item{abs_colour}{Logical. If TRUE, will use the absolute correlation for determining square colour.}

\item{set_colour_to_p_value}{Logical. If TRUE, will set the colour to p-value.}

\item{cluster_variables}{Logical. If TRUE, clusters the variables on the plot. This also moves data with no missing observations to the bottom.}

\item{n_decimal_places}{Number of decimal places to use for plotting.}

\item{suppress_warning_message}{Logical. If TRUE, suppresses warning message about the function being in beta.}
}
\value{
A ggplot object showing how missingness in variables is correlated with other variables.
}
\description{
This function plots how missingness in variables is correlated with other variables in the dataset.
}
\details{
This is a very early version of the function and requires additional testing and verification.
}
\examples{
data <- data.frame(A = rnorm(100), B = rnorm(100), C = rnorm(100))
data$A[sample(1:100, 20)] <- NA
data$B[sample(1:100, 30)] <- NA
plot_missing_correlations(data)

}
