% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairwise_missing.R
\name{plot_pairwise_missing}
\alias{plot_pairwise_missing}
\title{Create a ggplot2 plot of pairwise missingness}
\usage{
plot_pairwise_missing(
  dat,
  textadjust = 4,
  low_colour = "white",
  high_colour = "#0072B2",
  cluster_variables = FALSE,
  suppress_warning_message = FALSE,
  divisor = 1
)
}
\arguments{
\item{dat}{A dataframe or matrix for which to compute pairwise sample sizes.}

\item{textadjust}{A scalar that adjusts text size by a specific magnification factor.}

\item{low_colour}{Hex colour code for the lowest correlation value.}

\item{high_colour}{Hex colour code for the highest correlation value.}

\item{cluster_variables}{Logical. If TRUE, variables are clustered in the plot. Default is FALSE.}

\item{suppress_warning_message}{Logical. If TRUE, suppresses warning message about the function being in beta.}

\item{divisor}{A scalar that is divided to each sample size cell (default = 1). Helpful plotting for very large sample sizes.}
}
\value{
A ggplot object of the correlation matrix.
}
\description{
The function provides a wrapper around ggplot2 to easily create a paired sample size matrix
}
