% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_score.R
\name{sum_score}
\alias{sum_score}
\title{Calculate Sum Scores with Optional Mean Imputation}
\usage{
sum_score(
  input = NULL,
  max_percent_missing_allowed = 1,
  check_loadings = TRUE,
  plot_scores = FALSE,
  print_missing_table = FALSE,
  print_missing_each_input_var = FALSE,
  custom_score_calc = function(x) mean(x, na.rm = TRUE) * length(x)
)
}
\arguments{
\item{input}{A data frame or a numeric vector containing the variables for sum score calculation.
Ensure that only the relevant variables are included.}

\item{max_percent_missing_allowed}{Maximum allowed proportion of missing values for each observation.
Observations with missing data beyond this threshold are assigned `NA`.}

\item{check_loadings}{If `TRUE`, checks the loadings of the first principal component for
negative values, which may indicate data issues. For psychological questionnaire data,
this indicates that an item needs to be reverse coded.}

\item{plot_scores}{If `TRUE`, plots a histogram of the calculated sum scores.}

\item{print_missing_table}{If `TRUE`, prints a table summarizing the percentage of missing
data across all observations.}

\item{print_missing_each_input_var}{If `TRUE`, prints a table of missing data counts for each input variable.}

\item{custom_score_calc}{A custom function to calculate the sum score, defaults to mean imputation.}
}
\value{
A numeric vector of sum scores for each observation.
}
\description{
Calculates a sum score for each observation in a dataset, with handling for missing values.
The sum score is the product of the mean of non-missing values and the number of variables.
The function allows specifying the tolerance for missing data and provides additional diagnostic outputs.
}
\details{
The function operates row-wise. For rows with missing data within the permissible limit,
it replaces missing values with the mean of the available values in that row before calculating
the sum score. It can also perform a loading check of the first principal component,
plot the distribution of the scores, and provide detailed missing data diagnostics.
}
\examples{
# Assuming iris_test is a modified iris dataset with intentional NAs
iris_test = iris
iris_test[1,1:3] = NA
iris_test[c(4,15,16),2] = NA
sum_score(iris_test[,1:4], max_percent_missing_allowed = .2, print_missing_table = TRUE)

}
