% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlations.R
\name{plot_correlations}
\alias{plot_correlations}
\title{Create a ggplot2 Correlation Matrix}
\usage{
plot_correlations(
  dat,
  variable_labels = NULL,
  textadjust = 2,
  sample_size = TRUE,
  confidence_interval = TRUE,
  low_colour = "#0072B2",
  high_colour = "#D55E00",
  mid_colour = "white",
  abs_colour = TRUE,
  cluster_variables = FALSE,
  n_decimal_places = 2,
  suppress_warning_message = FALSE
)
}
\arguments{
\item{dat}{A dataframe or matrix for which to compute correlations. Note: A correlation matrix should not be provided as input.}

\item{variable_labels}{A character vector of labels for each variable (i.e., each column in `dat`). If missing (NULL), column names from `dat` are used.}

\item{textadjust}{A scalar that adjusts text size by a specific magnification factor.}

\item{sample_size}{Logical. If TRUE, the function includes sample sizes on the upper diagonal. If FALSE, these are left blank.}

\item{confidence_interval}{Logical. If TRUE, the function includes confidence intervals on the upper diagonal. If FALSE, these are left blank.}

\item{low_colour}{Hex colour code for the lowest correlation value.}

\item{high_colour}{Hex colour code for the highest correlation value.}

\item{mid_colour}{Hex colour code for the middle point in the correlation scale.}

\item{abs_colour}{Logical. If TRUE, absolute correlation values (i.e., ignoring positive/negative status) are used for determining square colours.}

\item{cluster_variables}{Logical. If TRUE, variables are clustered in the plot. Default is FALSE.}

\item{n_decimal_places}{The number of decimal places to use for plot text.}

\item{suppress_warning_message}{Logical. If TRUE, suppresses warning message about the function being in beta.}
}
\value{
A ggplot object of the correlation matrix.
}
\description{
The function provides a wrapper around ggplot2 to easily create a correlation matrix.
}
\examples{
X = sapply(1:10, function(i) rnorm(100))
X = as.data.frame(X)
My_Labels = c(paste0("Predictor ",1:5), paste0("Outcome ",1:5))

plot_correlations(X, variable_labels = My_Labels, sample_size = TRUE, confidence_interval = FALSE)

plot_correlations(X, variable_labels = My_Labels, sample_size = TRUE, confidence_interval = FALSE) +
ggplot2::labs(title = "My Title")
}
