% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Estimate Reliability from Bayesian statistical models}
\usage{
reliability(input_draws, verbose = TRUE)
}
\arguments{
\item{input_draws}{A matrix or data frame of posterior draws. Rows represent subjects and columns represent draws.}

\item{verbose}{A logical value indicating whether to print detailed information about the input data. Default is TRUE.}
}
\value{
A list with the following components:
\describe{
  \item{hdci}{A data frame with the highest density continuous interval (HDCI) for the reliability posterior.}
  \item{pd}{A numeric value representing the posterior probability of the direction of the reliability.}
  \item{reliability_posterior}{A numeric vector of reliability values for each posterior draw.}
  \item{input_draws_1}{A matrix of the first half of the shuffled posterior draws.}
  \item{input_draws_2}{A matrix of the second half of the shuffled posterior draws.}
}
}
\description{
This approach measures of the internal consistency reliability of posterior estimates from a Bayesian model.
}
\details{
For more detail, see our pre-print: ADD LINK


The function calculates the reliability by randomly splitting the posterior draws into two halves and then computing
the Pearson correlation between corresponding columns of the two halves. This process is repeated for each pair
of columns, and the resulting correlations form the reliability posterior distribution. The function then calculates
the 95% highest density continuous interval (HDCI) and the posterior probability of the direction (pd) of the reliability.
}
\examples{
\dontrun{
# Generate some example data
set.seed(123)
example_draws <- matrix(rnorm(1000), nrow = 50, ncol = 20)

# Calculate reliability
result <- reliability(example_draws, verbose = TRUE)
print(result)
}

}
\references{
ADD REF
}
